+++ src/index.test.ts
@@ -1,9 +1,18 @@
 import { describe, expect, test } from '@jest/globals'
 import { setTimeout } from 'node:timers/promises'
-import { defer, batchQueue, rateLimit, pausable, pacemaker } from './fns'
+import {
+  defer,
+  batchQueue,
+  rateLimit,
+  pausable,
+  pacemaker,
+  waitUntil,
+  TimeoutError,
+} from './fns'
 
 describe('rateLimit', () => {
   test('should add up to limit - 1 promises without delay', async () => {
+    expect.assertions(1)
     const limiter = rateLimit(3)
     const startTime = new Date().getTime()
     await limiter.add(setTimeout(1000))
@@ -13,6 +22,7 @@ describe('rateLimit', () => {
     expect(elapsed).toBeLessThan(100)
   })
   test('should wait for one promise to resolve when limit is reached', async () => {
+    expect.assertions(1)
     const limiter = rateLimit(3)
     const startTime = new Date().getTime()
     await limiter.add(setTimeout(1000))
@@ -23,6 +33,7 @@ describe('rateLimit', () => {
     expect(elapsed).toBeGreaterThan(900)
   })
   test('should finish awaiting remaining promises', async () => {
+    expect.assertions(3)
     const limiter = rateLimit(3)
     const done: string[] = []
     const createProm = () =>
@@ -54,6 +65,7 @@ describe('rateLimit', () => {
     await limiter.add(setTimeout(10))
   })
   test('rejections should bubble up .finish', () => {
+    expect.assertions(1)
     return expect(async () => {
       const limiter = rateLimit(3)
       await limiter.add(setTimeout(10))
@@ -70,9 +82,11 @@ describe('rateLimit', () => {
 
 describe('batchQueue', () => {
   test('should batch items up to batchSize', async () => {
+    expect.assertions(2)
     const calls: any[] = []
-    const fn = async (records: any[]) => {
+    const fn = async (records: string[]) => {
       calls.push(records)
+      return records.length
     }
     const batchSize = 2
 
@@ -84,10 +98,12 @@ describe('batchQueue', () => {
     }
     await queue.flush()
     expect(calls).toEqual([['Joe', 'Frank'], ['Bob']])
+    expect(queue.lastResult).toBe(1)
   })
   test('should flush queue if timeout is reached before batchSize', async () => {
+    expect.assertions(1)
     const calls: any[] = []
-    const fn = async (records: any[]) => {
+    const fn = async (records: string[]) => {
       calls.push(records)
     }
     const batchSize = 2
@@ -104,8 +120,9 @@ describe('batchQueue', () => {
     expect(calls).toEqual([['Joe'], ['Frank'], ['Bob']])
   })
   test('should reset timer if batchSize is reached before timeout', async () => {
+    expect.assertions(1)
     const calls: any[] = []
-    const fn = async (records: any[]) => {
+    const fn = async (records: string[]) => {
       calls.push(records)
     }
     const batchSize = 2
@@ -122,8 +139,9 @@ describe('batchQueue', () => {
     expect(calls).toEqual([['Joe', 'Frank'], ['Bob']])
   })
   test('fn should only be called once if timeout flush was triggered before queue.flush()', async () => {
+    expect.assertions(1)
     const calls: any[] = []
-    const fn = async (records: any[]) => {
+    const fn = async (records: string[]) => {
       calls.push(records)
       await setTimeout(100)
     }
@@ -141,8 +159,9 @@ describe('batchQueue', () => {
     expect(calls).toEqual([['Joe', 'Frank', 'Bob']])
   })
   test('should flush queue if batchBytes is reached before batchSize', async () => {
+    expect.assertions(1)
     const calls: any[] = []
-    const fn = async (records: any[]) => {
+    const fn = async (records: string[]) => {
       calls.push(records)
     }
     const batchSize = 3
@@ -161,8 +180,9 @@ describe('batchQueue', () => {
     ])
   })
   test('should flush queue if batchSize is reached before batchBytes', async () => {
+    expect.assertions(1)
     const calls: any[] = []
-    const fn = async (records: any[]) => {
+    const fn = async (records: string[]) => {
       calls.push(records)
     }
     const batchSize = 3
@@ -181,6 +201,7 @@ describe('batchQueue', () => {
 
 describe('pausable', () => {
   test('should pause and resume', async () => {
+    expect.assertions(2)
     const shouldProcess = pausable()
     const processed: string[] = []
     const processRecord = async (record: string) => {
@@ -202,6 +223,7 @@ describe('pausable', () => {
     expect(records).toEqual(['Joe', 'Frank', 'Bob'])
   })
   test('should pause and resume after timeout', async () => {
+    expect.assertions(2)
     const shouldProcess = pausable(50)
     const processed: string[] = []
     const processRecord = async (record: string) => {
@@ -225,6 +247,7 @@ describe('pausable', () => {
 
 describe('defer', () => {
   test('should defer', async () => {
+    expect.assertions(1)
     const delay = (milliseconds: number) => {
       const deferred = defer()
       global.setTimeout(deferred.done, milliseconds, '🦄')
@@ -240,6 +263,7 @@ describe('defer', () => {
 
 describe('pacemaker', () => {
   test('should call heartbeatFn 2 times', async () => {
+    expect.assertions(2)
     let count = 0
     const heartbeatFn = () => {
       console.log('heartbeat')
@@ -251,6 +275,7 @@ describe('pacemaker', () => {
   })
 
   test('should throw', async () => {
+    expect.assertions(1)
     const heartbeatFn = () => {
       console.log('heartbeat')
     }
@@ -261,5 +286,55 @@ describe('pacemaker', () => {
     }
 
     await expect(pacemaker(heartbeatFn, promFn(), 100)).rejects.toMatch('fail')
+  })
+})
+
+describe('waitUntil', () => {
+  test('should wait until pred returns truthy', async () => {
+    expect.assertions(1)
+    let isTruthy = false
+    global.setTimeout(() => {
+      isTruthy = true
+    }, 250)
+    await expect(waitUntil(() => isTruthy)).resolves.toBeUndefined()
+  })
+  test('should wait until async pred returns truthy', async () => {
+    expect.assertions(1)
+    let isTruthy = false
+    global.setTimeout(() => {
+      isTruthy = true
+    }, 250)
+    await expect(waitUntil(async () => isTruthy)).resolves.toBeUndefined()
+  })
+  test('should handle infinite timeout', async () => {
+    expect.assertions(1)
+    let isTruthy = false
+    global.setTimeout(() => {
+      isTruthy = true
+    }, 250)
+    await expect(
+      waitUntil(async () => isTruthy, { timeout: Infinity })
+    ).resolves.toBeUndefined()
+  })
+  test('should throw TimeoutError if the timeout expires', async () => {
+    expect.assertions(1)
+    let isTruthy = false
+    global.setTimeout(() => {
+      isTruthy = true
+    }, 250)
+    await expect(
+      waitUntil(async () => isTruthy, { timeout: 100 })
+    ).rejects.toThrow(new TimeoutError('Did not complete in 100 ms'))
+  })
+  test('rejects when pred throws', async () => {
+    expect.assertions(1)
+    await expect(
+      waitUntil(
+        async () => {
+          throw 'fail'
+        },
+        { timeout: 100 }
+      )
+    ).rejects.toMatch('fail')
   })
 })